feat: refactor code to multiple files
